+++ src/minigames/__init__.py
@@ -1,3 +1,19 @@
 """
 Minigames package.
 """
+
+from src.minigames.BaseMinigame import BaseMinigame
+from src.minigames.LockpickMinigame import LockpickMinigame
+from src.minigames.SafeMinigame import SafeMinigame
+from src.minigames.CrowbarMinigame import CrowbarMinigame
+from src.minigames.FuseBoxMinigame import FuseBoxMinigame
+from src.minigames.MinigameFactory import MinigameFactory
+
+__all__ = [
+    "BaseMinigame",
+    "LockpickMinigame",
+    "SafeMinigame",
+    "CrowbarMinigame",
+    "FuseBoxMinigame",
+    "MinigameFactory",
+]